Move send_user_event to account/utils.py
