+++ saleor/plugins/openid_connect/tests/test_utils.py
@@ -43,7 +43,6 @@ from ..utils import (
     get_user_from_oauth_access_token_in_jwt_format,
     get_user_from_token,
     get_user_info,
-    send_user_event,
     validate_refresh_token,
 )
 
@@ -1061,113 +1060,3 @@ def test_update_user_details_nothing_changed(
     customer_user.refresh_from_db()
     assert customer_user.first_name == first_name
     assert updated is False
-
-
-@patch("saleor.plugins.manager.PluginsManager.customer_updated")
-@patch("saleor.plugins.manager.PluginsManager.customer_created")
-@patch("saleor.plugins.manager.PluginsManager.staff_updated")
-@patch("saleor.plugins.manager.PluginsManager.staff_created")
-def test_send_user_event_no_webhook_sent(
-    mock_staff_created_webhook,
-    mock_staff_updated_webhook,
-    mock_customer_created_webhook,
-    mock_customer_updated_webhook,
-    customer_user,
-):
-    # when
-    send_user_event(customer_user, False, False)
-
-    # then
-    flush_post_commit_hooks()
-    mock_staff_created_webhook.assert_not_called()
-    mock_staff_updated_webhook.assert_not_called()
-    mock_customer_created_webhook.assert_not_called()
-    mock_customer_updated_webhook.assert_not_called()
-
-
-@patch("saleor.plugins.manager.PluginsManager.customer_updated")
-@patch("saleor.plugins.manager.PluginsManager.customer_created")
-@patch("saleor.plugins.manager.PluginsManager.staff_updated")
-@patch("saleor.plugins.manager.PluginsManager.staff_created")
-def test_send_user_event_customer_created_event(
-    mock_staff_created_webhook,
-    mock_staff_updated_webhook,
-    mock_customer_created_webhook,
-    mock_customer_updated_webhook,
-    customer_user,
-):
-    # when
-    send_user_event(customer_user, True, True)
-
-    # then
-    flush_post_commit_hooks()
-    mock_customer_created_webhook.assert_called_once_with(customer_user)
-    mock_customer_updated_webhook.assert_not_called()
-    mock_staff_created_webhook.assert_not_called()
-    mock_staff_updated_webhook.assert_not_called()
-
-
-@patch("saleor.plugins.manager.PluginsManager.customer_updated")
-@patch("saleor.plugins.manager.PluginsManager.customer_created")
-@patch("saleor.plugins.manager.PluginsManager.staff_updated")
-@patch("saleor.plugins.manager.PluginsManager.staff_created")
-def test_send_user_event_customer_updated_event(
-    mock_staff_created_webhook,
-    mock_staff_updated_webhook,
-    mock_customer_created_webhook,
-    mock_customer_updated_webhook,
-    customer_user,
-):
-    # when
-    send_user_event(customer_user, False, True)
-
-    # then
-    flush_post_commit_hooks()
-    mock_customer_updated_webhook.assert_called_once_with(customer_user)
-    mock_customer_created_webhook.assert_not_called()
-    mock_staff_created_webhook.assert_not_called()
-    mock_staff_updated_webhook.assert_not_called()
-
-
-@patch("saleor.plugins.manager.PluginsManager.customer_updated")
-@patch("saleor.plugins.manager.PluginsManager.customer_created")
-@patch("saleor.plugins.manager.PluginsManager.staff_updated")
-@patch("saleor.plugins.manager.PluginsManager.staff_created")
-def test_send_user_event_staff_created_event(
-    mock_staff_created_webhook,
-    mock_staff_updated_webhook,
-    mock_customer_created_webhook,
-    mock_customer_updated_webhook,
-    staff_user,
-):
-    # when
-    send_user_event(staff_user, True, True)
-
-    # then
-    flush_post_commit_hooks()
-    mock_staff_created_webhook.assert_called_once_with(staff_user)
-    mock_staff_updated_webhook.assert_not_called()
-    mock_customer_created_webhook.assert_not_called()
-    mock_customer_updated_webhook.assert_not_called()
-
-
-@patch("saleor.plugins.manager.PluginsManager.customer_updated")
-@patch("saleor.plugins.manager.PluginsManager.customer_created")
-@patch("saleor.plugins.manager.PluginsManager.staff_updated")
-@patch("saleor.plugins.manager.PluginsManager.staff_created")
-def test_send_user_event_staff_updated_event(
-    mock_staff_created_webhook,
-    mock_staff_updated_webhook,
-    mock_customer_created_webhook,
-    mock_customer_updated_webhook,
-    staff_user,
-):
-    # when
-    send_user_event(staff_user, False, True)
-
-    # then
-    flush_post_commit_hooks()
-    mock_staff_updated_webhook.assert_called_once_with(staff_user)
-    mock_staff_created_webhook.assert_not_called()
-    mock_customer_created_webhook.assert_not_called()
-    mock_customer_updated_webhook.assert_not_called()
